[!]Using modiInfos cache in BAIN refresh WIP - needs next: RRR
This should even further speedup the initial BAIN refresh at the cost of
missing some crc updates if moInfos isn't updated. That should be
extremely rare and a bug so should be fixed - but hence the !
I actually run into a case in the code that modinfos wasn't updated,
namely in Monitor External Install, so better safe than sorry.

Monitor install is kind of an exception to the rule that the data model
should take care of refreshes - it's a bit like we need to call
RefreshData (as if we tabbed out/in Bash). I added the needed refreshes

I address the rest of the cases I could find (after a lot of effort) in
the next commit - but it remains a fact that the cache for ModInfos (and
SaveInfos / BsaInfos) should also be indexed by ctime - since we
explicitly setmtime to the same value for various purposes. This is
under # 336 and we have some way there.

This is the reason that I did not entirely skip the special treatment of
mods in InstallersData#_process_data_dir if you were wondering (you
should have been). ModInfos is aware (modulo the next commit) of ctime,
so it will catch changes that _process_data_dir won't.

All this points to another refactoring idea i've been pondering on -
namely we need to refresh per directory not per infos singleton. As it
is we scan Data dir for mods and bsas _and_ on BAIN refresh, scan once.
That's under # 353

Under # 219.